Có add thêm folder images trong areas/content

+++ electronics-shop/Models/Product.cs
@@ -32,7 +32,7 @@ namespace electronics_shop.Models
         public Nullable<int> Quantity { get; set; }
         public string Description { get; set; }
         public Nullable<int> ViewCount { get; set; }
-        public Nullable<int> Rate { get; set; }
+        public Nullable<double> Rate { get; set; }
     
         public virtual Category Category { get; set; }
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
